moved some test functions to test_core
