+++ tests/test_mpdist.py
@@ -1,28 +1,11 @@
-import functools
-import math
-
 import naive
 import numpy as np
 import numpy.testing as npt
 import pytest
 from dask.distributed import Client, LocalCluster
 
-from stumpy import mpdist, mpdisted, stump
-from stumpy.core import _compute_P_ABBA, _mpdist
+from stumpy import mpdist, mpdisted
 from stumpy.mpdist import _mpdist_vect
-
-
-def some_func(P_ABBA, m, percentage, n_A, n_B):
-    percentage = min(percentage, 1.0)
-    percentage = max(percentage, 0.0)
-    k = min(math.ceil(percentage * (n_A + n_B)), n_A - m + 1 + n_B - m + 1 - 1)
-    P_ABBA.sort()
-    MPdist = P_ABBA[k]
-    if ~np.isfinite(MPdist):  # pragma: no cover
-        k = np.count_nonzero(np.isfinite(P_ABBA[:k])) - 1
-        MPdist = P_ABBA[k]
-
-    return MPdist
 
 
 @pytest.fixture(scope="module")
@@ -45,66 +28,6 @@ test_data = [
 
 percentage = [0.25, 0.5, 0.75]
 k = [0, 1, 2, 3, 4]
-
-
-@pytest.mark.parametrize("T_A, T_B", test_data)
-def test_compute_P_ABBA(T_A, T_B):
-    m = 3
-    n_A = T_A.shape[0]
-    n_B = T_B.shape[0]
-    ref_P_ABBA = np.empty(n_A - m + 1 + n_B - m + 1, dtype=np.float64)
-    comp_P_ABBA = np.empty(n_A - m + 1 + n_B - m + 1, dtype=np.float64)
-
-    ref_P_ABBA[: n_A - m + 1] = naive.stump(T_A, m, T_B)[:, 0]
-    ref_P_ABBA[n_A - m + 1 :] = naive.stump(T_B, m, T_A)[:, 0]
-
-    partial_stump = functools.partial(stump)
-    _compute_P_ABBA(T_A, T_B, m, comp_P_ABBA, partial_stump)
-
-    npt.assert_almost_equal(ref_P_ABBA, comp_P_ABBA)
-
-
-@pytest.mark.parametrize("T_A, T_B", test_data)
-def test_compute_P_ABBA_with_isconstant(T_A, T_B):
-    m = 3
-    n_A = T_A.shape[0]
-    n_B = T_B.shape[0]
-
-    T_A_subseq_isconstant = np.random.choice([True, False], n_A - m + 1, replace=True)
-    T_B_subseq_isconstant = np.random.choice([True, False], n_B - m + 1, replace=True)
-
-    ref_P_ABBA = np.empty(n_A - m + 1 + n_B - m + 1, dtype=np.float64)
-    comp_P_ABBA = np.empty(n_A - m + 1 + n_B - m + 1, dtype=np.float64)
-
-    ref_P_ABBA[: n_A - m + 1] = naive.stump(
-        T_A,
-        m,
-        T_B,
-        T_A_subseq_isconstant=T_A_subseq_isconstant,
-        T_B_subseq_isconstant=T_B_subseq_isconstant,
-    )[:, 0]
-    ref_P_ABBA[n_A - m + 1 :] = naive.stump(
-        T_B,
-        m,
-        T_A,
-        T_A_subseq_isconstant=T_B_subseq_isconstant,
-        T_B_subseq_isconstant=T_A_subseq_isconstant,
-    )[:, 0]
-
-    mp_func = functools.partial(
-        stump,
-        T_A_subseq_isconstant=T_A_subseq_isconstant,
-        T_B_subseq_isconstant=T_B_subseq_isconstant,
-    )
-    _compute_P_ABBA(
-        T_A,
-        T_B,
-        m,
-        comp_P_ABBA,
-        mp_func,
-    )
-
-    npt.assert_almost_equal(ref_P_ABBA, comp_P_ABBA)
 
 
 @pytest.mark.parametrize("T_A, T_B", test_data)
@@ -218,26 +141,6 @@ def test_mpdist_k(T_A, T_B, k):
     m = 3
     ref_mpdist = naive.mpdist(T_A, T_B, m, k=k)
     comp_mpdist = mpdist(T_A, T_B, m, k=k)
-
-    npt.assert_almost_equal(ref_mpdist, comp_mpdist)
-
-
-@pytest.mark.parametrize("T_A, T_B", test_data)
-@pytest.mark.parametrize("k", k)
-def test_mpdist_custom_func(T_A, T_B, k):
-    m = 3
-
-    percentage = 0.05
-    n_A = T_A.shape[0]
-    n_B = T_B.shape[0]
-
-    partial_k_func = functools.partial(
-        some_func, m=m, percentage=percentage, n_A=n_A, n_B=n_B
-    )
-    ref_mpdist = naive.mpdist(T_A, T_B, m)
-
-    partial_stump = functools.partial(stump)
-    comp_mpdist = _mpdist(T_A, T_B, m, partial_stump, custom_func=partial_k_func)
 
     npt.assert_almost_equal(ref_mpdist, comp_mpdist)
